[6X] gpinitsystem should return code 1 when invalid option is used
Backport of https://github.com/greenplum-db/gpdb/pull/16857/

Issue: gpinitsystem fails with an error When an invalid option
is provided. But the exit code is 0 not 1. It should fail with
exit code 1.

Fix: LOG_MSG fn is being used to print the error which sets
EXIT_STATUS to 1 only when message is a FATAL or WARNING message.
Updated LOG_MSG to set the exit_status as 1 in case of ERROR messages also.

Added behave scenario to validate the changes